updated work order to use reusable ui components
